rewrote methods, renamed
